requirements.txt
django
python-dotenv
djangorestframework
psycopg2-binary
drf-writable-nested